+++ tests/test_ElectromechanicalPneumaticILC.cpp
@@ -74,8 +74,8 @@ protected:
 
     void processResetServer(uint8_t address) override {}
 
-    void processHardpointForceStatus(uint8_t address, uint8_t status, int32_t encoderPosition,
-                                     float loadCellForce) override {}
+    void processStepperForceStatus(uint8_t address, uint8_t status, int32_t encoderPosition,
+                                   float loadCellForce) override {}
 
     void processDCAGain(uint8_t address, float primaryGain, float secondaryGain) override {}