+++ doc/version-history.rst
@@ -8,6 +8,7 @@ v1.10.1
 * Set and read DCA Gain commands (ILC code 73 and 74)
 * Add NiWarning class, improve Ni error reporting
 * FPGA communication debugging
+* support for M2 command line client
 
 v1.10.0
 -------